refactor: colocate link, markdown, preview, row, section, text in react-email
Add marked to react-email dependencies; external marked for markdown package builds.

Made-with: Cursor
