+++ packages/markdown/package.json
@@ -22,8 +22,8 @@
     }
   },
   "scripts": {
-    "build": "tsdown src/index.ts --format esm,cjs --dts --external react",
-    "build:watch": "tsdown src/index.ts --format esm,cjs --dts --external react --watch",
+    "build": "tsdown ../react-email/src/components/markdown/index.ts --format esm,cjs --dts --external react --external marked --outDir dist",
+    "build:watch": "tsdown ../react-email/src/components/markdown/index.ts --format esm,cjs --dts --external react --external marked --outDir dist --watch",
     "clean": "rm -rf dist",
     "test": "vitest run",
     "test:watch": "vitest"
