+++ packages/react-email/src/components/index.ts
@@ -10,3 +10,9 @@ export * from './heading/index.js';
 export * from './hr/index.js';
 export * from './html/index.js';
 export * from './img/index.js';
+export * from './link/index.js';
+export * from './markdown/index.js';
+export * from './preview/index.js';
+export * from './row/index.js';
+export * from './section/index.js';
+export * from './text/index.js';
