+++ packages/react-email/src/components/link/index.ts
@@ -1,0 +1,1 @@
+export * from './link.js';
