+++ packages/react-email/src/components/markdown/index.ts
@@ -1,0 +1,1 @@
+export * from './markdown.js';
