+++ packages/react-email/src/components/preview/index.ts
@@ -1,0 +1,1 @@
+export * from './preview.js';
