+++ packages/react-email/src/components/section/index.ts
@@ -1,0 +1,1 @@
+export * from './section.js';
